Game/Back Game/requirements.txt
mysql-connector-python == 8.0.33
mysqlclient == 2.1.1
Flask == 2.3.2